@@ -1,67 +1,66 @@
 {
-    "name": "webapp",
-    "version": "0.0.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "^7.2.15",
-        "@angular/cdk": "^7.3.7",
-        "@angular/common": "^7.2.15",
-        "@angular/compiler": "^7.2.15",
-        "@angular/core": "^7.2.15",
-        "@angular/flex-layout": "^7.0.0-beta.24",
-        "@angular/forms": "^7.2.15",
-        "@angular/http": "^7.2.15",
-        "@angular/material": "^7.3.7",
-        "@angular/material-moment-adapter": "^7.3.7",
-        "@angular/platform-browser": "^7.2.15",
-        "@angular/platform-browser-dynamic": "^7.2.15",
-        "@angular/router": "^7.2.15",
-        "@swimlane/ngx-charts": "^9.0.0",
-        "@types/webpack-env": "^1.13.6",
-        "angular-calendar": "^0.25.2",
-        "core-js": "^2.6.1",
-        "d3": "^5.6.0",
-        "date-fns": "^1.30.1",
-        "faker": "^4.1.0",
-        "hammerjs": "^2.0.8",
-        "moment": "^2.24.0",
-        "ngx-mask": "^8.0.6",
-        "rxjs": "^6.4.0",
-        "ts-md5": "^1.2.4",
-        "tslib": "^1.9.0",
-        "zone.js": "~0.8.29"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "^0.803.14",
-        "@angular/cli": "~7.3.9",
-        "@angular/compiler-cli": "^7.2.15",
-        "@angular/language-service": "^7.2.15",
-        "@angularclass/hmr": "^2.1.3",
-        "@types/faker": "^4.1.5",
-        "@types/jasmine": "~3.3.13",
-        "@types/jasminewd2": "~2.0.6",
-        "@types/node": "^11.13.4",
-        "codelyzer": "~5.0.0",
-        "coveralls": "^3.0.2",
-        "firebase-tools": "^6.2.2",
-        "jasmine-core": "~2.99.1",
-        "jasmine-spec-reporter": "~4.2.1",
-        "karma": "^4.3.0",
-        "karma-chrome-launcher": "~2.2.0",
-        "karma-coverage-istanbul-reporter": "~2.0.1",
-        "karma-jasmine": "~1.1.2",
-        "karma-jasmine-html-reporter": "^0.2.2",
-        "ng-bullet": "^1.0.3",
-        "ts-node": "~8.2.0",
-        "tslint": "~5.16.0",
-        "typescript": "~3.1.6"
-    }
+  "name": "webapp",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^8.2.12",
+    "@angular/cdk": "^7.3.7",
+    "@angular/common": "^8.2.12",
+    "@angular/compiler": "^8.2.12",
+    "@angular/core": "^8.2.12",
+    "@angular/flex-layout": "^7.0.0-beta.24",
+    "@angular/forms": "^8.2.12",
+    "@angular/material": "^7.3.7",
+    "@angular/material-moment-adapter": "^7.3.7",
+    "@angular/platform-browser": "^8.2.12",
+    "@angular/platform-browser-dynamic": "^8.2.12",
+    "@angular/router": "^8.2.12",
+    "@swimlane/ngx-charts": "^12.0.1",
+    "ngx-mask": "^8.0.6",
+    "@types/webpack-env": "^1.13.6",
+    "angular-calendar": "^0.25.2",
+    "core-js": "^2.6.1",
+    "d3": "^5.6.0",
+    "date-fns": "^1.30.1",
+    "faker": "^4.1.0",
+    "hammerjs": "^2.0.8",
+    "moment": "^2.24.0",
+    "rxjs": "^6.4.0",
+    "ts-md5": "^1.2.4",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.9.1"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.803.14",
+    "@angular/cli": "~8.3.14",
+    "@angular/compiler-cli": "^8.2.12",
+    "@angular/language-service": "^8.2.12",
+    "@angularclass/hmr": "^2.1.3",
+    "@types/faker": "^4.1.5",
+    "@types/jasmine": "~3.3.13",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/node": "^11.13.4",
+    "codelyzer": "^5.0.1",
+    "coveralls": "^3.0.2",
+    "firebase-tools": "^6.2.2",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "^4.3.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "ng-bullet": "^1.0.3",
+    "ts-node": "~8.2.0",
+    "tslint": "~5.16.0",
+    "typescript": "~3.5.3"
+  }
 }
